Tweaking the DB Car images separate table
Placing car Images in a separate table.

What is done:
-Frontpage scaling bug fixed.
- README updated somewhat.
- New Local DB (after format)
