+++ README.md
@@ -1,3 +1,2 @@
-Readme file.
-
-This project is based on Google App Engines Django Skeleton.
+Car Rental Website Project.
+This project is based on Google App Engines Django Skeleton. Follow this guide for the setup: [GAE Django Skeleton Github](https://github.com/GoogleCloudPlatform/appengine-django-skeleton)
