+++ frontpage/templates/frontpage/index.html
@@ -54,13 +54,13 @@
                 <div class="row">
                      <div class="col-md-12 col-sm-12 col-xs-12">
                         <h3>Biltype</h3>
-                        <div class="checkbox-inline">
+                        <div class="checkbox-inline col-md-1 col-sm-12 col-xs-12">
                           <label><input type="checkbox" name="personal" id="personal">Personbil</label>
                         </div>
-                        <div class="checkbox-inline">
+                        <div class="checkbox-inline col-md-1 col-sm-12 col-xs-12">
                           <label><input type="checkbox" name="van" id="van">Varebil</label>
                         </div>
-                        <div class="checkbox-inline">
+                        <div class="checkbox-inline col-md-1 col-sm-12 col-xs-12">
                           <label><input type="checkbox" name="combi_car" id="combi_car">Kombinertbil</label>
                         </div>
                      </div>
