+++ requirements-vendor.txt
@@ -1,1 +1,5 @@
 Django
+reportlab
+
+
+# pip install -r requirements-vendor.txt -t lib
